requirements.txt
torch>=2.0.0
transformers>=4.37.0
datasets
accelerate
peft
bitsandbytes
pillow
numpy
tqdm
pyyaml
scikit-learn
matplotlib
seaborn
